pyproject.toml

[tool.poetry]
name = "discord-security-bot"
version = "0.1.0"
description = "Bot de sécurité Discord"
authors = ["Replit User"]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
